Add abstract ObjectEventSortingInput
